+++ script/build.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+bun compile 
+exit 1