+++ CHANGELOG.md
@@ -1,8 +1,30 @@
 # Changelog
 
+## 3.0.0
+
+### Changed
+- Convert project into a GitHub CLI extension
+
+### Added
+
+- Support for notifications (`-n` or `--notify`)
+  - Notification when new PR is added
+  - Notification when one of your PRs becomes mergable
+  - Notification when one of your PRs fails CI/CD
+  - Notification when one of your PRs gets a new comment
+
+### Fixed
+
+- Don't block main thread to make exiting the script easier
+
+### Removed
+
+- Remove `--sound` flag in favour of `--notify`.
+
 ## 2.0.0
 
-### Added 
+### Added
+
 - Show number of comments.
 - Show when PR has conflicts.
 - Show when PR is mergable.
@@ -12,33 +34,40 @@
 - Support for multiple repositories (`--repos` replaces `--repo`).
 
 ### Fixed
+
 - Don't play sound on first run.
 - Stop waiting for beeper to resolve causing tiny lag.
 
 ### Removed
+
 - Remove `--repo` flag in favour of `--repos`.
 
 ## 1.0.4 - 2024-01-12
 
 ### Added
+
 - Added `--involved` flag.
 
 ### Changed
+
 - Print output in alternate buffer.
 
 ## 1.0.3 - 2023-12-13
 
 ### Fixed
+
 - Fix formatting when review decision is missing.
 
 ## 1.0.2 - 2023-12-01
 
 ### Fixed
+
 - Fallback to username for authors missing name.
 
 ## 1.0.1 - 2023-11-17
 
 ### Added
+
 - `--labels` flag for only showing pull requests that needs review from you with any of the specified labels.
 - Show failing checks with links.
 - Show additions and deletions.
@@ -46,4 +75,5 @@
 - Loading indicator while updating pull requests.
 
 ### Fixed
+
 - Error handling when not authenticated to github.